specify run explicitly and limit reruns to 1

+++ .github/workflows/check_repo_ci_jobs.py
@@ -8,6 +8,7 @@ import os
 REPO_OWNER = os.environ.get("REPO_OWNER", "krande/condapackaging")
 MY_WORKFLOW = os.environ.get("MY_WORKFLOW", "ci-force-fail")
 TOKEN = os.environ.get("GITHUB_TOKEN", None)
+WORKFLOW_RUN_ID = os.environ.get("WORKFLOW_RUN_ID", None)
 
 SIGNS_OF_STOPPAGE = [
     "Error: The operation was canceled.",
@@ -30,12 +31,11 @@ def start_request_session():
     return s
 
 
-def get_ci_jobs(repo_name):
-    url = f"https://api.github.com/repos/{repo_name}/actions/runs"
+def get_ci_run(repo_name, run_id):
+    url = f"https://api.github.com/repos/{repo_name}/actions/runs/{run_id}"
     s = start_request_session()
     response = s.get(url)
-    jobs = response.json()
-    return list(filter(lambda x: x["name"] == MY_WORKFLOW, jobs["workflow_runs"]))
+    return response.json()
 
 
 def evaluate_log_file_for_abrupt_stop(log_file):
@@ -77,44 +77,19 @@ def restart_job(repo_name, job_id):
     return response.json()
 
 
-def check_logs_for_status_code_143(logs):
-    """Do something like this?"""
-    for log in logs["lines"]:
-        if log["name"] == "stderr":
-            if "Error: The operation was canceled." in log["text"]:
-                return True
-            if "fatal error C1060: compiler is out of heap space" in log["text"]:
-                return True
-    return False
-
-
-def get_failed_jobs():
-    jobs = get_ci_jobs(REPO_OWNER)
-    failed_jobs = []
-    for job in filter(lambda x: x["name"] == MY_WORKFLOW, jobs):
-        if job["name"] != MY_WORKFLOW:
-            continue
-
-        if job["status"] != "completed":
-            continue
-        if job["conclusion"] != "failure":
-            continue
-
-        failed_jobs.append(job)
-    failed_jobs = list(sorted(failed_jobs, key=lambda x: x["run_number"]))
-    return failed_jobs
-
-
 def eval_jobs():
-    failed_jobs = get_failed_jobs()
+    run = get_ci_run(REPO_OWNER, WORKFLOW_RUN_ID)
+    if run["run_attempt"] > 1:
+        print("This is not the first attempt. Exiting")
+        return
+    failed_logs = get_ci_specific_run_specific_failure_details(REPO_OWNER, WORKFLOW_RUN_ID)
 
-    last_job = failed_jobs[-1]
-    logs = get_ci_specific_run_specific_failure_details(REPO_OWNER, last_job["id"])
+    if len(failed_logs) == 0:
+        print("No runs exhibit signs of abrupt stoppage")
+        return
 
-    print(logs)
-    print("restarting job", last_job["id"])
-
-    r = restart_job(REPO_OWNER, last_job["id"])
+    print("restarting job", WORKFLOW_RUN_ID)
+    r = restart_job(REPO_OWNER, WORKFLOW_RUN_ID)
     print(r)
 
     shutil.rmtree("logs")
